split-driver(01) [driver][NFC] Detach conv module
