+++ driver/main.cpp
@@ -26,9 +26,10 @@
 #include <iostream>
 #include <cstdio>
 
+#include "conv.hpp"
+
 #include "activ_driver.hpp"
 #include "bn_driver.hpp"
-#include "conv_driver.hpp"
 #include "CBAInferFusion_driver.hpp"
 #include "driver.hpp"
 #include "groupnorm_driver.hpp"
@@ -69,198 +70,177 @@ int main(int argc, char* argv[])
         std::cout << " " << argv[i];
     std::cout << std::endl;
 
-    Driver* drv;
-    if(base_arg == "conv")
+    Driver* drv = makeDriverConv(base_arg);
+    if(drv != nullptr)
     {
-        drv = new ConvDriver<float, float>();
-    }
-    else if(base_arg == "convfp16")
-    {
-        drv = new ConvDriver<float16, float>();
-    }
-    else if(base_arg == "convbfp16")
-    {
-        drv = new ConvDriver<bfloat16, float>();
-    }
-    else if(base_arg == "convint8")
-    {
-        drv = new ConvDriver<int8_t, int32_t>();
-    }
-    else if(base_arg == "convfp8")
-    {
-        drv = new ConvDriver<float8, float>();
-    }
-    else if(base_arg == "convbfp8")
-    {
-        drv = new ConvDriver<bfloat8, float>();
-    }
-    else if(base_arg == "CBAInfer")
-    {
-        drv = new CBAInferFusionDriver<float, double>();
-    }
-    else if(base_arg == "CBAInferfp16")
-    {
-        drv = new CBAInferFusionDriver<float16, double>();
-    }
-    else if(base_arg == "pool")
-    {
-        drv = new PoolDriver<float, double>();
-    }
-    else if(base_arg == "poolfp16")
-    {
-        drv = new PoolDriver<float16, double>();
-    }
-    else if(base_arg == "lrn")
-    {
-        drv = new LRNDriver<float, double>();
-    }
-    else if(base_arg == "lrnfp16")
-    {
-        drv = new LRNDriver<float16, double>();
-    }
-    else if(base_arg == "activ")
-    {
-        drv = new ActivationDriver<float, double>();
-    }
-    else if(base_arg == "activfp16")
-    {
-        drv = new ActivationDriver<float16, double>();
-    }
-    else if(base_arg == "softmax")
-    {
-        drv = new SoftmaxDriver<float, double>();
-    }
-    else if(base_arg == "softmaxfp16")
-    {
-        drv = new SoftmaxDriver<float16, double>();
-    }
+        if(base_arg == "CBAInfer")
+        {
+            drv = new CBAInferFusionDriver<float, double>();
+        }
+        else if(base_arg == "CBAInferfp16")
+        {
+            drv = new CBAInferFusionDriver<float16, double>();
+        }
+        else if(base_arg == "pool")
+        {
+            drv = new PoolDriver<float, double>();
+        }
+        else if(base_arg == "poolfp16")
+        {
+            drv = new PoolDriver<float16, double>();
+        }
+        else if(base_arg == "lrn")
+        {
+            drv = new LRNDriver<float, double>();
+        }
+        else if(base_arg == "lrnfp16")
+        {
+            drv = new LRNDriver<float16, double>();
+        }
+        else if(base_arg == "activ")
+        {
+            drv = new ActivationDriver<float, double>();
+        }
+        else if(base_arg == "activfp16")
+        {
+            drv = new ActivationDriver<float16, double>();
+        }
+        else if(base_arg == "softmax")
+        {
+            drv = new SoftmaxDriver<float, double>();
+        }
+        else if(base_arg == "softmaxfp16")
+        {
+            drv = new SoftmaxDriver<float16, double>();
+        }
 #if MIOPEN_USE_GEMM
-    else if(base_arg == "gemm")
-    {
-        drv = new GemmDriver<float>();
-    }
-    else if(base_arg == "gemmfp16")
-    {
-        drv = new GemmDriver<float16>();
-    }
+        else if(base_arg == "gemm")
+        {
+            drv = new GemmDriver<float>();
+        }
+        else if(base_arg == "gemmfp16")
+        {
+            drv = new GemmDriver<float16>();
+        }
 #endif
-    else if(base_arg == "bnorm")
-    {
-        drv = new BatchNormDriver<float, double>();
+        else if(base_arg == "bnorm")
+        {
+            drv = new BatchNormDriver<float, double>();
+        }
+        else if(base_arg == "bnormfp16")
+        {
+            drv = new BatchNormDriver<float16, double, float>();
+        }
+        else if(base_arg == "rnn_seq")
+        {
+            drv = new RNNSeqDriver<float, double>();
+        }
+        else if(base_arg == "rnn_seqfp16")
+        {
+            drv = new RNNSeqDriver<float16, double>();
+        }
+        else if(base_arg == "rnn")
+        {
+            drv = new RNNDriver<float, double>();
+        }
+        else if(base_arg == "rnnfp16")
+        {
+            drv = new RNNDriver<float16, double>();
+        }
+        else if(base_arg == "ctc")
+        {
+            drv = new CTCDriver<float>();
+        }
+        else if(base_arg == "dropout")
+        {
+            drv = new DropoutDriver<float, float>();
+        }
+        else if(base_arg == "dropoutfp16")
+        {
+            drv = new DropoutDriver<float16, float>();
+        }
+        else if(base_arg == "groupnorm")
+        {
+            drv = new GroupNormDriver<float, double>();
+        }
+        else if(base_arg == "groupnormfp16")
+        {
+            drv = new GroupNormDriver<float16, double>();
+        }
+        else if(base_arg == "groupnormbfp16")
+        {
+            drv = new GroupNormDriver<bfloat16, double>();
+        }
+        else if(base_arg == "tensorop")
+        {
+            drv = new TensorOpDriver<float, float>();
+        }
+        else if(base_arg == "tensoropfp16")
+        {
+            drv = new TensorOpDriver<float16, float>();
+        }
+        else if(base_arg == "reduce")
+        {
+            drv = new ReduceDriver<float, float>();
+        }
+        else if(base_arg == "reducefp16")
+        {
+            drv = new ReduceDriver<float16, float>();
+        }
+        else if(base_arg == "reducefp64")
+        {
+            drv = new ReduceDriver<double, double>();
+        }
+        else if(base_arg == "layernorm")
+        {
+            drv = new LayerNormDriver<float, float>();
+        }
+        else if(base_arg == "layernormfp16")
+        {
+            drv = new LayerNormDriver<float16, float>();
+        }
+        else if(base_arg == "layernormbfp16")
+        {
+            drv = new LayerNormDriver<bfloat16, float>();
+        }
+        else if(base_arg == "sum")
+        {
+            drv = new SumDriver<float, float>();
+        }
+        else if(base_arg == "sumfp16")
+        {
+            drv = new SumDriver<float16, float>();
+        }
+        else if(base_arg == "sumbfp16")
+        {
+            drv = new SumDriver<bfloat16, float>();
+        }
+        else if(base_arg == "argmax")
+        {
+            drv = new ArgmaxDriver<float, float>();
+        }
+        else if(base_arg == "argmaxfp16")
+        {
+            drv = new ArgmaxDriver<float16, float>();
+        }
+        else if(base_arg == "argmaxbfp16")
+        {
+            drv = new ArgmaxDriver<bfloat16, float>();
+        }
+        else if(base_arg == "cat")
+        {
+            drv = new CatDriver<float>();
+        }
+        else if(base_arg == "catfp16")
+        {
+            drv = new CatDriver<float16>();
+        }
+        else if(base_arg == "catbfp16")
+        {
+            drv = new CatDriver<bfloat16>();
+        }
     }
-    else if(base_arg == "bnormfp16")
-    {
-        drv = new BatchNormDriver<float16, double, float>();
-    }
-    else if(base_arg == "rnn_seq")
-    {
-        drv = new RNNSeqDriver<float, double>();
-    }
-    else if(base_arg == "rnn_seqfp16")
-    {
-        drv = new RNNSeqDriver<float16, double>();
-    }
-    else if(base_arg == "rnn")
-    {
-        drv = new RNNDriver<float, double>();
-    }
-    else if(base_arg == "rnnfp16")
-    {
-        drv = new RNNDriver<float16, double>();
-    }
-    else if(base_arg == "ctc")
-    {
-        drv = new CTCDriver<float>();
-    }
-    else if(base_arg == "dropout")
-    {
-        drv = new DropoutDriver<float, float>();
-    }
-    else if(base_arg == "dropoutfp16")
-    {
-        drv = new DropoutDriver<float16, float>();
-    }
-    else if(base_arg == "groupnorm")
-    {
-        drv = new GroupNormDriver<float, double>();
-    }
-    else if(base_arg == "groupnormfp16")
-    {
-        drv = new GroupNormDriver<float16, double>();
-    }
-    else if(base_arg == "groupnormbfp16")
-    {
-        drv = new GroupNormDriver<bfloat16, double>();
-    }
-    else if(base_arg == "tensorop")
-    {
-        drv = new TensorOpDriver<float, float>();
-    }
-    else if(base_arg == "tensoropfp16")
-    {
-        drv = new TensorOpDriver<float16, float>();
-    }
-    else if(base_arg == "reduce")
-    {
-        drv = new ReduceDriver<float, float>();
-    }
-    else if(base_arg == "reducefp16")
-    {
-        drv = new ReduceDriver<float16, float>();
-    }
-    else if(base_arg == "reducefp64")
-    {
-        drv = new ReduceDriver<double, double>();
-    }
-    else if(base_arg == "layernorm")
-    {
-        drv = new LayerNormDriver<float, float>();
-    }
-    else if(base_arg == "layernormfp16")
-    {
-        drv = new LayerNormDriver<float16, float>();
-    }
-    else if(base_arg == "layernormbfp16")
-    {
-        drv = new LayerNormDriver<bfloat16, float>();
-    }
-    else if(base_arg == "sum")
-    {
-        drv = new SumDriver<float, float>();
-    }
-    else if(base_arg == "sumfp16")
-    {
-        drv = new SumDriver<float16, float>();
-    }
-    else if(base_arg == "sumbfp16")
-    {
-        drv = new SumDriver<bfloat16, float>();
-    }
-    else if(base_arg == "argmax")
-    {
-        drv = new ArgmaxDriver<float, float>();
-    }
-    else if(base_arg == "argmaxfp16")
-    {
-        drv = new ArgmaxDriver<float16, float>();
-    }
-    else if(base_arg == "argmaxbfp16")
-    {
-        drv = new ArgmaxDriver<bfloat16, float>();
-    }
-    else if(base_arg == "cat")
-    {
-        drv = new CatDriver<float>();
-    }
-    else if(base_arg == "catfp16")
-    {
-        drv = new CatDriver<float16>();
-    }
-    else if(base_arg == "catbfp16")
-    {
-        drv = new CatDriver<bfloat16>();
-    }
-    else
+    if(drv == nullptr)
     {
         printf("Incorrect BaseArg\n");
         exit(0); // NOLINT (concurrency-mt-unsafe)
